WIP: flesh out `wradlib.georef` submodules, clean-up zonalstats, fix tests/notebooks,
